refactor: course list and empty state
